Minor changes and a new method
