+++ game_board.rb
@@ -1,0 +1,30 @@
+require 'terminal-table'
+
+class GameBoard
+  def initialize
+
+  end
+
+  def success(good_letters_locs) # needs the array of elements including letters & index
+    puts
+  end
+
+  def print_game_board(current_letter)
+    puts "Amazing answer. You are guessing: #{current_letter}"
+  end
+
+def dash_line
+  dash_line = ""
+
+# dash.each do |letter|
+#   dash_line = dash_line + "#{ dash[i]} "
+
+end
+end
+
+end #GameBoard
+
+# my_word_guess = WordGuess.new
+# word_guess = my_word_guess.word_logic
+# my_game_board = GameBoard.new
+# my_game_board.print_game_board(word_guess)
